fix : data type integrate(Intager -> Long)

+++ backend/shopping-mall-api/src/main/java/com/shopping_mall_api/dto/cart/CartResponseDTO.java
@@ -16,7 +16,7 @@ import java.util.List;
 public class CartResponseDTO {
     private Long userId;
     private List<CartItemResponseDTO> cartItemList;
-    private Integer totalCartPrice;
+    private Long totalCartPrice;
 
     public CartResponseDTO(Cart cart){
         if(cart == null){
